Refactor startup classes
