+++ src/Speakr.WebApp.Site/AppStart/Configuration.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.Extensions.Configuration;
+
+namespace Speakr.WebApp.Site.AppStart
+{
+    public class Configuration
+    {
+        public static IConfigurationRoot Configure()
+        {
+            var builder = new ConfigurationBuilder();
+
+            return builder.Build();
+        }
+    }
+}
