+++ src/Speakr.WebApp.Site/Program.cs
@@ -1,5 +1,6 @@
 ﻿using System.IO;
 using Microsoft.AspNetCore.Hosting;
+using Speakr.WebApp.AppStart;
 
 namespace Speakr.WebApp
 {
