--- src/Speakr.WebApp.Site/appsettings.json
@@ -1,7 +1,0 @@
-﻿{
-    "DefaultHealthString": "AppSettings Default Healthy Response",
-
-    "ConnectionStrings": {
-        "DefaultConnection": "Server=(localdb)\\MSSQLLocalDB;Database=_CHANGE_ME;Trusted_Connection=True;MultipleActiveResultSets=true"
-    }
-}
